+++ parallel-based-sorting-algorithms/src/main/java/michaeljosh/brickSort.java
@@ -3,30 +3,30 @@ package michaeljosh;
 import java.util.Arrays;
 
 public class brickSort {
-    public static void brickSort(int arr[]) {
-        int n = arr.length;
+    public static void brickSort(long arr[]) {
+        long n = arr.length;
         boolean isSorted = false; // Initially array is unsorted
 
         while (!isSorted) {
             isSorted = true;
-            int temp = 0;
+            long temp = 0;
 
             // Perform Bubble sort on odd indexed element
-            for (int i = 1; i <= n - 2; i = i + 2) {
-                if (arr[i] > arr[i + 1]) {
-                    temp = arr[i];
-                    arr[i] = arr[i + 1];
-                    arr[i + 1] = temp;
+            for (long i = 1; i <= n - 2; i = i + 2) {
+                if (arr[(int)i] > arr[(int)(i + 1)]) {
+                    temp = arr[(int)i];
+                    arr[(int)i] = arr[(int)(i + 1)];
+                    arr[(int)(i + 1)] = temp;
                     isSorted = false;
                 }
             }
 
             // Perform Bubble sort on even indexed element
-            for (int i = 0; i <= n - 2; i = i + 2) {
-                if (arr[i] > arr[i + 1]) {
-                    temp = arr[i];
-                    arr[i] = arr[i + 1];
-                    arr[i + 1] = temp;
+            for (long i = 0; i <= n - 2; i = i + 2) {
+                if (arr[(int)i] > arr[(int)(i + 1)]) {
+                    temp = arr[(int)i];
+                    arr[(int)i] = arr[(int)(i + 1)];
+                    arr[(int)(i + 1)] = temp;
                     isSorted = false;
                 }
             }
@@ -34,7 +34,7 @@ public class brickSort {
     }
 
     public static void main(String[] args) {
-        int arr[] = {34, 2, 10, -9};
+        long arr[] = {34, 2, 10, -9};
         brickSort(arr);
         System.out.println(Arrays.toString(arr));
     }
